+++ src/Variables.cpp
@@ -2,6 +2,7 @@
 
 int Variables::DailyLeft = 0;
 int Variables::WeeklyLeft = 0;
+int Variables::EventLeft = 0;
 
 int Variables::GlobalRank = 0;
 int Variables::OldStarsCount = 0;
@@ -9,3 +10,4 @@ int Variables::OldStarsCount = 0;
 cocos2d::SEL_MenuHandler Variables::BISelector = nullptr;
 cocos2d::SEL_MenuHandler Variables::SupExSelector = nullptr;
 cocos2d::SEL_MenuHandler Variables::GDDPSelector = nullptr;
+cocos2d::SEL_MenuHandler Variables::RouletteSelector = nullptr;