+++ src/Variables.hpp
@@ -4,6 +4,7 @@ using namespace geode::prelude;
 namespace Variables {
     extern int DailyLeft;
     extern int WeeklyLeft;
+    extern int EventLeft;
 
     extern int GlobalRank;
     extern int OldStarsCount;
@@ -11,4 +12,5 @@ namespace Variables {
     extern cocos2d::SEL_MenuHandler BISelector;
     extern cocos2d::SEL_MenuHandler SupExSelector;
     extern cocos2d::SEL_MenuHandler GDDPSelector;
+    extern cocos2d::SEL_MenuHandler RouletteSelector;
 }