+++ src/hooks/LayersHooks.hpp
@@ -4,6 +4,11 @@
 #include "../Variables.hpp"
 using namespace geode::prelude;
 
+#define getSelector(id, var) \
+	if (auto button = creatorLayer->getChildByIDRecursive(id)) {\
+		var = as<CCMenuItemSpriteExtra*>(button)->m_pfnSelector;\
+	}
+
 std::vector<int> mainLevels = { 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 5001, 5002, 5003, 5004 };
 int thing = 0;
 
@@ -29,69 +34,12 @@ class $modify(LoadingLayer) {
 		auto loader = Loader::get();
         if (thing == 8) {
             auto creatorLayer = CreatorLayer::create(); // phantom layer rip
-			
-			if (loader->isModLoaded("cvolton.betterinfo")) {
-				if (loader->isModLoaded("undefined0.gdtweaks")) {
-					if (loader->getLoadedMod("undefined0.gdtweaks")->getSettingValue<bool>("replace-more-games-w-texture")) {
-						if (auto menu = creatorLayer->getChildByID("bottom-left-menu")) {
-							if (auto button = menu->getChildByID("cvolton.betterinfo/main-button")) {
-								Variables::BISelector = as<CCMenuItemSpriteExtra*>(button)->m_pfnSelector;
-							}
-						}
-					} else {
-						if (auto menu = creatorLayer->getChildByID("cvolton.betterinfo/center-right-menu")) {
-							if (auto button = menu->getChildByID("cvolton.betterinfo/main-button")) {
-								Variables::BISelector = as<CCMenuItemSpriteExtra*>(button)->m_pfnSelector;
-							}
-						}
-					}
-				} else {
-					if (auto menu = creatorLayer->getChildByID("cvolton.betterinfo/center-right-menu")) {
-						if (auto button = menu->getChildByID("cvolton.betterinfo/main-button")) {
-							Variables::BISelector = as<CCMenuItemSpriteExtra*>(button)->m_pfnSelector;
-						}
-					}
-				}
-			}
 
-			auto pagesMod = loader->getLoadedMod("alphalaneous.pages_api");
-			if (pagesMod->getSettingValue<bool>("creator-layer-menu")) {
-				if (auto pageMenu = creatorLayer->getChildByID("paged-creator-buttons-menu")) {
-					if (auto node = pageMenu->getChildByID("pages")) {
-						CCArrayExt<CCMenu*> pages = node->getChildren();
-						for (auto& page : pages) {
-							CCArrayExt<CCNode*> buttons = page->getChildren();
-							for (auto& button : buttons) {
-								if (button->getID() == "super-expert-button" && loader->isModLoaded("xanii.super_expert")) {
-									Variables::SupExSelector = as<CCMenuItemSpriteExtra*>(button)->m_pfnSelector;
-								} else if (button->getID() == "demon-progression-button" && loader->isModLoaded("minemaker0430.gddp_integration")) {
-									Variables::GDDPSelector = as<CCMenuItemSpriteExtra*>(button)->m_pfnSelector;
-								}
-							}
-						}
-					}
-				}
-			} else {
-				if (auto menu = creatorLayer->getChildByID("creator-buttons-menu")) {
-					if (loader->isModLoaded("xanii.super_expert")) {
-						if (auto button = menu->getChildByID("super-expert-button")) {
-							Variables::SupExSelector = as<CCMenuItemSpriteExtra*>(button)->m_pfnSelector;
-						}
-					}
-					if (loader->isModLoaded("minemaker0430.gddp_integration")) {
-						if (auto button = menu->getChildByID("demon-progression-button")) {
-							Variables::GDDPSelector = as<CCMenuItemSpriteExtra*>(button)->m_pfnSelector;
-						}
-					}
-				}
-			}
-
-			// i hate gdutils
-			if (loader->isModLoaded("gdutilsdevs.gdutils")) {
-				if (loader->getLoadedMod("gdutilsdevs.gdutils")->getSettingValue<bool>("activate-background")) {
-					creatorLayer->retain();
-				}
-			}
+			// thanks minecraftify
+			getSelector("cvolton.betterinfo/main-button", Variables::BISelector);
+			getSelector("spaghettdev.gd-roulette/roulette-button", Variables::RouletteSelector);
+			getSelector("super-expert-button", Variables::SupExSelector);
+			getSelector("demon-progression-button", Variables::GDDPSelector);
         }
     }
 };